+++ Lib/test/libregrtest/findtests.py
@@ -25,10 +25,11 @@ SPLITTESTDIRS: set[TestName] = {
     "test_gdb",
     "test_inspect",
     "test_io",
-    "test_pydoc",
     "test_multiprocessing_fork",
     "test_multiprocessing_forkserver",
     "test_multiprocessing_spawn",
+    "test_os",
+    "test_pydoc",
 }
