+++ Lib/test/test_os/utils.py
@@ -1,0 +1,3 @@
+def create_file(filename, content=b'content'):
+    with open(filename, "xb", 0) as fp:
+        fp.write(content)
